third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-script-no-cors-no-xorigin-empty-intergity.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS disabled script, without CORS fetch, but an empty integrity attribute</title>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        setup({single_test: true});

        var result = '';
        var scriptLoaded = function() {
            assert_equals(result, 'ran', 'Script ran.');
            done();
        };
        var scriptError = function() {
            assert_unreached('Script did not run.');
            done();
        };
    </script>
    <script integrity="" src="http://localhost:8000/security/resources/cors-script.php?cors=false&value=ran" onload="scriptLoaded();" onerror="scriptError()"></script>
</body>
</html>
