third_party/blink/web_tests/platform/win11-arm64/http/tests/security/subresourceIntegrity/subresource-integrity-script-no-cors-no-xorigin-empty-intergity-expected.txt

CONSOLE ERROR: Uncaught Error: assert_equals: Script ran. expected "ran" but got ""
This is a testharness.js-based test.
FAIL SRI on cross origin CORS disabled script, without CORS fetch, but an empty integrity attribute Uncaught Error: assert_equals: Script ran. expected "ran" but got ""
Harness: the test ran to completion.
